feat: added tasks module, moved api module to more relevant folder
